api/requirements.txt
flask>=3.0
supabase>=2.0
sendgrid>=6.0
python-dotenv>=1.0
